api-gateway/pyproject.toml
[project]
name = "api-gateway"
version = "0.1.0"
requires-python = ">=3.10"

dependencies = [
    "fastapi",
    "uvicorn",
    "aio-pika",
    "python-jose",
    "tenacity",
    "aiobreaker",
    "prometheus-client",
    "opentelemetry-sdk",
    "opentelemetry-api",
    "opentelemetry-instrumentation-fastapi",
]

[project.optional-dependencies]
dev = ["watchfiles"]


[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]